refactor: 관심사 분리를 위해 query client를 분리했지만 이는 devtool이 겹쳐서 보이는 문제와 cache 관점에서 비효율적이라 판단해 최상위 query client 하나만 사용하도록 변경
