+++ src/app/(post)/layout.tsx
@@ -1,15 +1,8 @@
 import { ReactNode } from 'react';
-import {
-  CheckLoginSessionProvider,
-  LoginSessionRQProvider,
-} from '@/components/context';
+import { CheckLoginSessionProvider } from '@/components/context';
 
 export default function Layout({
   children,
 }: Readonly<{ children: ReactNode }>) {
-  return (
-    <LoginSessionRQProvider>
-      <CheckLoginSessionProvider>{children}</CheckLoginSessionProvider>
-    </LoginSessionRQProvider>
-  );
+  return <CheckLoginSessionProvider>{children}</CheckLoginSessionProvider>;
 }
