--- src/components/context/LoginSessionRQProvider/index.tsx
@@ -1,44 +1,0 @@
-'use client';
-
-import {
-  QueryCache,
-  QueryClient,
-  QueryClientProvider,
-} from '@tanstack/react-query';
-import { useRouter } from 'next/navigation';
-import { ReactNode, useState } from 'react';
-import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
-import CheckLoginSessionError from '@/services/errors/CheckLoginSessionError';
-
-export default function LoginSessionRQProvider({
-  children,
-}: Readonly<{ children: ReactNode }>) {
-  const router = useRouter();
-  const [queryClient] = useState(() => {
-    return new QueryClient({
-      defaultOptions: {
-        queries: {
-          refetchOnWindowFocus: false,
-          retry: false,
-        },
-      },
-      queryCache: new QueryCache({
-        onError: (error) => {
-          if (error instanceof CheckLoginSessionError) {
-            alert('재 로그인이 필요합니다.');
-            router.push('/login');
-          }
-        },
-      }),
-    });
-  });
-
-  return (
-    <QueryClientProvider client={queryClient}>
-      {children}
-      <ReactQueryDevtools
-        initialIsOpen={process.env.NEXT_PUBLIC_MODE === 'local'}
-      />
-    </QueryClientProvider>
-  );
-}
